+++ gamemodes/terrortown/entities/entities/ttt_c4/shared.lua
@@ -418,7 +418,7 @@ if SERVER then
 	---
 	-- @realm server
 	function ENT:OnRemove()
-		markerVision.RemoveEntity(self)
+		self:RemoveMarkerVision("c4_owner")
 	end
 
 	---
@@ -436,7 +436,7 @@ if SERVER then
 		self:SetExplodeTime(0)
 		self:SetArmed(false)
 
-		markerVision.RemoveEntity(self)
+		self:RemoveMarkerVision("c4_owner")
 
 		self.DisarmCausedExplosion = false
 	end
@@ -508,7 +508,10 @@ if SERVER then
 
 		events.Trigger(EVENT_C4PLANT, ply)
 
-		markerVision.RegisterEntity(self, ply, VISIBLE_FOR_TEAM)
+		local mvObject = self:AddMarkerVision("c4_owner")
+		mvObject:SetOwner(ply)
+		mvObject:SetVisibleFor(VISIBLE_FOR_TEAM)
+		mvObject:SyncToClients()
 	end
 
 	---
@@ -825,8 +828,9 @@ else -- CLIENT
 
 	hook.Add("TTT2RenderMarkerVisionInfo", "HUDDrawMarkerVisionC4", function(mvData)
 		local ent = mvData:GetEntity()
+		local mvObject = mvData:GetMarkerVisionObject()
 
-		if not IsValid(ent) or ent:GetClass() ~= "ttt_c4" then return end
+		if not mvObject:IsObjectFor(ent, "c4_owner") then return end
 
 		local owner = ent:GetOriginator()
 		local nick = IsValid(owner) and owner:Nick() or "---"
@@ -842,7 +846,7 @@ else -- CLIENT
 		mvData:AddDescriptionLine(ParT("c4_marker_vision_time", {time = time}))
 		mvData:AddDescriptionLine(ParT("marker_vision_distance", {distance = distance}))
 
-		mvData:AddDescriptionLine(TryT("marker_vision_visible_for_" .. markerVision.GetVisibleFor(ent)), COLOR_SLATEGRAY)
+		mvData:AddDescriptionLine(TryT(mvObject:GetVisibleForTranslationKey()), COLOR_SLATEGRAY)
 
 		local color = COLOR_WHITE