+++ gamemodes/terrortown/entities/entities/ttt_radio.lua
@@ -36,7 +36,10 @@ function ENT:Initialize()
 	if SERVER then
 		self:SetUseType(SIMPLE_USE)
 
-		markerVision.RegisterEntity(self, self:GetOriginator(), VISIBLE_FOR_TEAM)
+		local mvObject = self:AddMarkerVision("radio_owner")
+		mvObject:SetOwner(self:GetOwner())
+		mvObject:SetVisibleFor(VISIBLE_FOR_TEAM)
+		mvObject:SyncToClients()
 	end
 
 	-- Register with owner
@@ -97,7 +100,7 @@ function ENT:OnRemove()
 
 		client.radio = nil
 	else
-		markerVision.RemoveEntity(self)
+		self:RemoveMarkerVision("radio_owner")
 	end
 end
 ---
@@ -320,8 +323,9 @@ if CLIENT then
 
 	hook.Add("TTT2RenderMarkerVisionInfo", "HUDDrawMarkerVisionRadio", function(mvData)
 		local ent = mvData:GetEntity()
+		local mvObject = mvData:GetMarkerVisionObject()
 
-		if not IsValid(ent) or ent:GetClass() ~= "ttt_radio" then return end
+		if not mvObject:IsObjectFor(ent, "radio_owner") then return end
 
 		local originator = ent:GetOriginator()
 		local nick = IsValid(originator) and originator:Nick() or "---"
@@ -336,7 +340,7 @@ if CLIENT then
 		mvData:AddDescriptionLine(ParT("marker_vision_owner", {owner = nick}))
 		mvData:AddDescriptionLine(ParT("marker_vision_distance", {distance = distance}))
 
-		mvData:AddDescriptionLine(TryT("marker_vision_visible_for_" .. markerVision.GetVisibleFor(ent)), COLOR_SLATEGRAY)
+		mvData:AddDescriptionLine(TryT(mvObject:GetVisibleForTranslationKey()), COLOR_SLATEGRAY)
 	end)
 end