+++ gamemodes/terrortown/gamemode/server/sv_main.lua
@@ -14,6 +14,7 @@ ttt_include("sh_door")
 ttt_include("sh_voice")
 ttt_include("sh_printmessage_override")
 ttt_include("sh_speed")
+ttt_include("sh_marker_vision_element")
 
 ttt_include("sv_network_sync")
 ttt_include("sv_hud_manager")