+++ lua/terrortown/lang/en.lua
@@ -2084,9 +2084,6 @@ L.flaregun_help_primary = "Burn body/entity"
 -- 2023-12-14
 L.marker_vision_owner = "Owner: {owner}"
 L.marker_vision_distance = "Distance: {distance}m"
-L.marker_vision_visible_for_0 = "Visible for: you"
-L.marker_vision_visible_for_1 = "Visible for: team"
-L.marker_vision_visible_for_2 = "Visible for: everyone"
 L.marker_vision_distance_collapsed = "{distance}m"
 
 L.c4_marker_vision_time = "Detonation time: {time}"
@@ -2170,6 +2167,12 @@ Dynamic crosshair enables scaling the crosshair depending on the weapon's cone. 
 If the line length is kept static, only the gap scales with cone changes.]]
 
 L.header_weapon_settings = "Weapon Settings"
+
+--2024-01-29
+L.marker_vision_visible_for_0 = "Visible for you"
+L.marker_vision_visible_for_1 = "Visible for your role"
+L.marker_vision_visible_for_2 = "Visible for your team"
+L.marker_vision_visible_for_3 = "Visible for everyone"
 
 -- 2024-01-27
 L.decoy_help_primary = "Throw Decoy on the ground"